fix(gitlab-ci): some globally defined keywords are deprecated
As defined in the documentation[1], defining globally `image` and
`cache` is deprecated. It should at least go under the `default`
keyword.

We define the `.meteor` job template to factorise all the meteor jobs
common attributes, like a meteor only `default`.

This is a preparation for futur non meteor jobs which will not require
most of them.

Footnotes:
[1]  https://docs.gitlab.com/ee/ci/yaml/#globally-defined-image-services-cache-before_script-after_script